Object occurrence validation

* added validation with GetById where it was possible

* covered every put and delete with validation

* one line format


Approved-by: Marcin D

+++ Schedule.API/Controllers/ParticipantController.cs
@@ -45,10 +45,11 @@ public class ParticipantController : ControllerBase
 	{
 		Participant? participant = await _participantService
 			.GetByIdAsync(participantId, companyId);
+		if (participant == null)
+			return NotFound();
 
 		_mapper.Map(request, participant);
-
-		await _participantService.PutAsync(participant!);
+		await _participantService.PutAsync(participant);
 		return NoContent();
 	}
 
@@ -58,6 +59,11 @@ public class ParticipantController : ControllerBase
 		Guid participantId
 	)
 	{
+		Participant? participant = await _participantService
+			.GetByIdAsync(participantId, companyId);
+		if (participant == null)
+			return NotFound();
+
 		await _participantService.DeleteByIdAsync(participantId, companyId);
 		return NoContent();
 	}
